refactor: 메서드명 및 패키지 이동
